refactor: rename parserError to parseError
