+++ src/index.ts
@@ -17,4 +17,4 @@ export * from './lib/trinosql/TrinoSqlVisitor';
 export { SyntaxContextType } from './parser/common/basic-parser-types';
 
 export type * from './parser/common/basic-parser-types';
-export type { SyntaxError, ParserError } from './parser/common/parserErrorListener';
+export type { SyntaxError, ParseError } from './parser/common/parseErrorListener';
